+++ src/vue/modules/create-sale-form/components/information-step-form.vue
@@ -391,6 +391,9 @@ export default {
     },
 
     populateForm () {
+      const quoteAssets = this.request.quoteAssets
+        .filter(quoteAsset => this.request.defaultQuoteAsset !== quoteAsset)
+
       this.form.name = this.request.name
       this.form.baseAsset = this.ownedAssets
         .find(item => item.code === this.request.baseAsset)
@@ -401,7 +404,7 @@ export default {
       this.form.softCap = this.request.softCap
       this.form.hardCap = this.request.hardCap
       this.form.assetsToSell = this.request.assetsToSell
-      this.form.quoteAssets = this.request.quoteAssets
+      this.form.quoteAssets = quoteAssets
       this.form.isWhitelisted = this.request.isWhitelisted
     },